Support for "reconnecting" Supervisor Comms from task process when `dag.test()` are used
This is a follow up to #57212, which worked fine "at run time" but did not
work in many of our own unit tests, which rely on `dag.test` or `ti.run`.

The way this is implemented is that when we use the InProcessTestSupervisor we
pre-emptively create a socket pair. We have to create it even it its not being
used, as we can't know.

And since this is all in one process we create a thread to handle the socket
comms. Since this is only ever for tests performance or hitting the GIL
doesn't matter.